+++ minchin/jrnl/plugins/base.py
@@ -8,6 +8,7 @@ Base class for Importers and Exporters.
 
 
 import os
+from pathlib import Path
 import re
 import unicodedata
 
@@ -53,25 +54,26 @@ class BaseExporter:
     @classmethod
     def write_file(cls, journal, path):
         """Exports a journal into a single file."""
+        path.parent.mkdir(parents=True, exist_ok=True)
         try:
             with open(path, "w", encoding="utf-8") as f:
                 f.write(cls.export_journal(journal))
-                return (
-                    f"[Journal '{journal.name}' exported (as a single file) to '{path}']"
-                )
+                return f"[Journal '{journal.name}' exported (as a single file) to '{path}']"
         except IOError as e:
-            return f"[{ERROR_COLOR}ERROR{RESET_COLOR}: {e.filename} {e.strerror}]"
+            return f"[{ERROR_COLOR}ERROR{RESET_COLOR}] '{e.filename}': {e.strerror}"
         except NotImplementedError:
             return (
-                f"[{ERROR_COLOR}ERROR{RESET_COLOR}: This exporter doesn't support "
-                "exporting as a single file.]"
+                f"[{ERROR_COLOR}ERROR{RESET_COLOR}] This exporter doesn't support "
+                "exporting as a single file. Try exporting to a directory instead?"
             )
 
     @classmethod
     def make_filename(cls, entry):
-        return entry.date.strftime("%Y-%m-%d") + "_{}.{}".format(
+        """Determine the filename to save an individual entry as."""
+        fn = entry.date.strftime("%Y-%m-%d") + "_{}.{}".format(
             cls._slugify(str(entry.title)), cls.extension
         )
+        return fn
 
     @classmethod
     def write_files(cls, journal, path):
@@ -79,20 +81,26 @@ class BaseExporter:
         try:
             for entry in journal.entries:
                 try:
-                    full_path = os.path.join(path, cls.make_filename(entry))
+                    base_path = Path(path)
+                    fn = cls.make_filename(entry)
+                    full_path = base_path / fn
+                    full_path.parent.mkdir(parents=True, exist_ok=True)
                     with open(full_path, "w", encoding="utf-8") as f:
                         f.write(cls.export_entry(entry))
                 except IOError as e:
-                    return "[{2}ERROR{3}: {0} {1}]".format(
+                    return "[{2}ERROR{3}] '{0}': {1}".format(
                         e.filename, e.strerror, ERROR_COLOR, RESET_COLOR
                     )
         except NotImplementedError:
             return (
-                f"[{ERROR_COLOR}ERROR{RESET_COLOR}: This exporter doesn't support "
-                "exporting as individual files.]"
+                f"[{ERROR_COLOR}ERROR{RESET_COLOR}] This exporter doesn't "
+                "support exporting as individual files. Try exporting to a "
+                "single file instead?"
             )
         else:
-            return f"[Journal '{journal.name}' exported (as multiple files) to '{path}']"
+            return (
+                f"[Journal '{journal.name}' exported (as multiple files) to '{path}']"
+            )
 
     def _slugify(string):
         """